Update siteorigin-panels.php
Argh, my mistake. I had written this filter for an older version of panels, and realized when I updated to the latest that it was referencing a $panels_data which was no longer in scope in this version. I simply removed the reference to the $panels_data and now it works fine.

Apologies.