Object cache the rendered content
Currently whenever the content or excerpt is called for a post, the
siteprigin plugin has to render the full widget content (and also
enqueue all the scripts and styles for the widgets). This is _very_ slow
(somewhere in the 300ms range, depending on plenty of things of course.)
Given the post content will not change until the post it's self is
updated, it makes sense to make use of the Object Cache API.

It looks like there was some in-memory cache, however this never
functioned, as it appears that there was only code to check the cache,
never populate it.

Because of the scripts and styles that can go with widgets, we have to
use a hacky complex solution to capture all the enqueued scripts and
styles while the widgets are being rendered. Unfortunatly this means
hacking with the globals, but I don't know of a better way to acheive
this other than a more declerative API for widget scripts and styles.

The patch was tested and singificantly improved page render time in my
testing (from about 800ms to 400ms) however this is going to depend a
lot on what widgets are used for example.