Cell Vertical Alignment: Only Use Required Selector
This change will ensure only the needed Cell Vertical Alignment selector is output. Previously, both the styled and unstyled selector was output but only one was ever output.

It'll also resolve a display issue with the CSS Container Breaker and stretched columns and a cell with a background applied but no widgets.